sstable: create sstable component files in a subdirectory
When writing the sstable, create a temporary directory
for creating all components so that each sstable files' will be
assigned a different allocaton groups on xfs.

Files are immediately renamed to their default location after creation.
Temp directory is removed when the sstable is sealed.

Additional work to be introduced in the following patches:
- When populating tables, temp directories need to be looked up and removed.

Fixes #3167

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>