sstables: continuous_data_consumer: Make position() meaningful inside state_processor::process_state()
Will allow state_processor to know its position in the
stream.

Currently position() is meaningless inside process_state() because in
some cases it points to the position after the buffer and in some
cases before it. This patch standardizes on the former. This is more
useful than the latter because process_state() trims from the front of
the buffer as it consumes, so the position inside the stream can be
obtained by subtracting the remaining buffer size from position(),
without introducing any new variables.